refactor to better structure
