+++ src/router.rs
@@ -1,12 +1,24 @@
+//! Router definitions and utilities
+//!
+//! This module contains:
+//! - Route definitions using ruled_router derive macros
+//! - Navigation helper functions
+//! - Browser history management
+//! - Popstate event handling
+
+use respo::DispatchFn;
 use ruled_router::error::RouteState;
 use ruled_router::prelude::*;
 use ruled_router_derive::{QueryDerive, RouterData, RouterMatch};
 use serde::{Deserialize, Serialize};
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsCast;
 
-use respo::DispatchFn;
-
-use crate::router_util::navigate_to;
 use crate::store::ActionOp;
+
+// ============================================================================
+// Route Definitions
+// ============================================================================
 
 /// Top-level route matcher (enum for matching different routes)
 /// Note: Order matters! More specific routes should come first.
@@ -47,17 +59,16 @@ pub struct CounterModuleRoute {
   pub sub_router: RouteState<CounterSubRouterMatch>,
 }
 
-// Serialize as URL string
 impl Serialize for CounterModuleRoute {
   fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
   where
     S: serde::Serializer,
   {
+    // Serialize as URL string for simpler storage format
     serializer.serialize_str(&self.format())
   }
 }
 
-// Deserialize from URL string
 impl<'de> Deserialize<'de> for CounterModuleRoute {
   fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
   where
@@ -103,7 +114,7 @@ pub struct CounterQuery {
 }
 
 // ============================================================================
-// App-specific navigation helpers
+// Navigation Helpers
 // ============================================================================
 
 /// Navigate to home page
@@ -128,4 +139,67 @@ pub fn navigate_counter_detail(dispatch: &DispatchFn<ActionOp>, id: u32) -> Resu
       })),
     }),
   )
+}
+
+/// Navigate to a new route by dispatching an action
+/// This follows the unidirectional data flow pattern:
+/// 1. Format route to URL path
+/// 2. Push to browser history
+/// 3. Dispatch action to update store
+pub fn navigate_to(dispatch: &DispatchFn<ActionOp>, route: AppRouterMatch) -> Result<(), String> {
+  let path = route.format();
+  push_history_state(&path);
+  dispatch.run(ActionOp::RouteChange(route))
+}
+
+// ============================================================================
+// Browser History Management
+// ============================================================================
+
+/// Push a new state to browser history
+fn push_history_state(path: &str) {
+  if let Some(window) = web_sys::window() {
+    if let Ok(history) = window.history() {
+      let _ = history.push_state_with_url(&JsValue::NULL, "", Some(path));
+    }
+  }
+}
+
+/// Get current route from browser URL
+pub fn get_current_route() -> AppRouterMatch {
+  web_sys::window()
+    .and_then(|w| {
+      let pathname = w.location().pathname().ok()?;
+      let search = w.location().search().unwrap_or_default();
+      let full_path = format!("{pathname}{search}");
+      Some(parse_route(&full_path))
+    })
+    .unwrap_or_default()
+}
+
+/// Parse URL path and return matched route
+fn parse_route(path: &str) -> AppRouterMatch {
+  AppRouterMatch::try_parse(path).unwrap_or_default()
+}
+
+// ============================================================================
+// Popstate Event Handling
+// ============================================================================
+
+/// Setup popstate event listener to handle browser back/forward navigation
+pub fn setup_router_listener<F>(on_route_change: F)
+where
+  F: Fn(AppRouterMatch) + 'static,
+{
+  let callback = Closure::wrap(Box::new(move |_event: web_sys::PopStateEvent| {
+    let route = get_current_route();
+    on_route_change(route);
+  }) as Box<dyn Fn(_)>);
+
+  if let Some(window) = web_sys::window() {
+    let _ = window.add_event_listener_with_callback("popstate", callback.as_ref().unchecked_ref());
+  }
+
+  // Leak the closure to keep it alive for the lifetime of the app
+  callback.forget();
 }
